Docs; tinkering
The Savon::Client @wsdl attribute stores useful info about each SOAP
method, including what parameters (and their types) are expected by each
method. They *appear* to be defined in order (which is important); that
is, #params_for_method returns a hash with the keys in the same order as
the Cherwell API documentation indicates, i.e.:

    > cherwell.client.params_for_method(:update_business_object)
    => {
      :busObNameOrId => {:name => "busObNameOrId", :type => "string"},
      :busObRecId => {:name => "busObRecId", :type => "string"},
      :updateXml => {:name => "updateXml", :type => "string"}
    }

matches:

    string UpdateBusinessObject(
      string busObNameOrId,
      string busObRecId,
      string updateXml
    )

Hoping to use this to advantage, allowing more direct invocation of Cherwell's
API methods without having to explicitly provide parameter names. That is,
instead of this:

    > cherwell.client.update_business_object(
        :busObNameOrId => "Incident",
        :busObRecId => "12345",
        :updateXml => "<?xml ...>"
      )

I'd like to support:

    > cherwell.client.update_business_object(
        "Incident", "12345", "<?xml ...>"
      )

Or, ideally, both.